requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
frozenlist==1.8.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
multidict==6.7.0
namex==0.1.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.1
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.5
rich==14.2.0
scikit-learn==1.6.1
scipy==1.13.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
twilio==9.8.6
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
yarl==1.22.0